Add `--max-comments-per-command` configuration
Allow Atlantis administrators to cap the number of VCS comments that
will be created for a single command. (The default value is 0, which
means unlimited comments.)

If you're trying something like `TF_LOG=debug`, Atlantis can produce so
many comments that it becomes challenging to read your PR, or worse
still, your VCS might rate-limit Atlantis, effectively breaking your
ability to use it for an extended period of time. While this PR does not
change the default behavior, it's probably a good idea to set this flag
to something like 10.

Fixes #416.