Update using-atlantis.md
I was having problems with `atlantis apply` when I wanted to apply all pending plans.

The problem was that every time I tried to execute this action, I got "Ran Apply for 0 projects:" because atlantis, for the apply action, did not know where the tfstates were. 

I think this is something that we can add in the documentation, because I did not find this clarification
The problem was that every time I tried to execute this action, I was getting "Ran Apply for 0 projects:" due to atlantis, for the apply action, didn’t know where the tfstates were located. 

I think it's something we can add in the documentation, because I didn’t find this aclaration