Fix rendering of :put_litter_in_its_place: in default templates
It seems that GitHub has changed its Markdown renderer such that an
emoji followed by a string of plain text (with no other formatting)
followed by a colon at the end of the line renders as plain text instead
of as the emoji. Compare the following two lines:

* :put_litter_in_its_place: To delete all plans and locks for the PR, comment:

* :put_litter_in_its_place: To **delete** all plans and locks for the PR, comment:

While this seems like it is probably a GitHub bug, we can work around it
by adding some formatting between the emoji and the end of the line,
which matches the bullet point above it anyway.