Allow for plans to be partially successful
When automerge is true, if any plan fails, all plans get deleted

This is true even when running from a github comment.
When dealing with many projects, there is a high likelihood that one
plan can fail.

This allows us to keep apply some plans and reiterate so we don't lose
all progress.