gitlab: link latest commit to the correct pipeline
PR #4785 was a partial fix, as it might link a commit to the wrong pipeline when
GitLab is slow to update the HeadPipeline.

This patch ensures the latest commit is always linked to the correct
pipeline by querying the pipeline associated with the commit's ref,
instead of relying on HeadPipeline.

Tested in our environment and it looks good so far.

Signed-off-by: Raul Gutierrez Segales <rsegales@nvidia.com>