Add new draftplan command.
draftplan acts like a normal plan command, except it does not grab a
project lock or a terraform lock, and does not refresh state.
The resulting plan will only be printed as a comment, it will
not be stored and so cannot be applied.

It is intended a a lightweight method to see what a plan would
look like.