apply gh app's URL fix on MergeAgain() too MergeAgain() ends up calling recheckDiverged() which reset remotes URLs using CloneURL.
Github App workingdir needs to fix the CloneURL of both remotes. Until recently this logic was done one Clone(). But recently MergeAgain() logic has been moved out of Clone().
So the same fix needs to be applied there.

Without this fix, recheckDiverged() fails when running `git remote update`. Because it will use empty credentials.
Also this failure is done silently because it does not return an error.

Looking at FileWorkspace methods, we potentially also want to run the same fix for `HasDiverged()` which runs `git fetch`. But for now this function does not reset the remotes URLs the same way recheckDiverged() does.

Signed-off-by: francois poinsot <francois.poinsot@cognite.com>