feat: VCS pull request status retries and backoff
Introduce retries and backoff to VCS pull request status checks that are done through
API calls of underlying provider.

When user issues an atlantis command, it is usually the case in huge repositories that atlantis has
not yet processed them all internally and it results in failure to perform a command.

Exposing retries and backoff will give atlantis server administrators power to make a concious
tradeoff between the speed of returning server responses and probability of returning failed status check
for better user experience which varies based on the repository size and it's traffic.

Signed-off-by: Krystian Kulgawczuk <150012886+krystian-kulgawczuk-natwest@users.noreply.github.com>