fix: add bounds checking for custom policy sets
When using custom policy checks, ensure we don't access inputPolicySets
out of bounds when iterating over outputs. If there are more outputs
than policy sets, fallback to 'Custom' as the policy set name.

This prevents potential index out of range panics when the number of
outputs exceeds the number of configured policy sets.

Signed-off-by: Pepe Amengual <pepe.amengual@example.com>
Signed-off-by: PePe Amengual <2208324+jamengual@users.noreply.github.com>
Signed-off-by: Brett Galkowski <bgalkows@uci.edu>
Signed-off-by: dimisjim <dimitris.moraitidis@gmail.com>